chore(ios): remove functions other than the minimum required
