+++ ios/Plugin/GooglePay/GooglePayExecutor.swift
@@ -1,0 +1,9 @@
+//
+//  GooglePayExecutor.swift
+//  Plugin
+//
+//  Created by 榊原昌彦 on 2021/07/11.
+//  Copyright © 2021 Max Lynch. All rights reserved.
+//
+
+import Foundation
