--- ios/Plugin/STPAuthCtx.swift
@@ -1,8 +1,0 @@
-import Foundation
-import Stripe
-
-extension StripePlugin: STPAuthenticationContext {
-    public func authenticationPresentingViewController() -> UIViewController {
-        return (self.bridge?.viewController)! as UIViewController
-    }
-}
